Expose CrmSvcUtilMetadataGenerator to allow clients to generate their own metadata more easily
